Dataset To be analysed
The Dataset has been prepared for analysing with all 'Yes' answers being represented by 1 and 'No' Answers being represented by 2. For Gender 'Male' is represented by 1 and 'Female' is represented by 0. Finally, for the 'class' section which identifies if the patient reported a positive or negative diabetes result, the 'Positive' result is represented by a 1 and a 'Negative' result is represented by a 0.

All ages have been divided by 100.